Use the Content-Length header if it is set
The Content-Length header is never used by the Go runtime.
Instead, the ContentLength field on the request struct is used to set the header
when writing the actual request. This field is only set for certain types of io.Readers.
Azure storage signatures include the Content-Length header, hence, it needs to be
respected, even if the io.Reader is some other type that is not recognized by the Go
runtime.

Signed-off-by: Paul Meyer <paul.meyer@microsoft.com>