storage: return AzureStorageServiceError for empty responses
Currently, if somebody was trying to implement an is404(err)
for errors returned from "storage" package, they would miss
out the errors coming from HEAD requests, such as GetBlobProperties
as these responses do not have a response body that contains an error
message.

Therefore, reusing AzureStorageServiceError to return a sensible
error message that still contains the status code (int) to implement
a err.(AzureStorageServiceError).Code==404 and the status text (str)
that contains a helpful error message instead of just "Not Found" with
an indicator saying "no response body was available".

That simplifies the code that would implement an is404(error) in the
downstream clients.

Signed-off-by: Ahmet Alp Balkan <ahmetalpbalkan@gmail.com>